requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic-settings==2.3.4
python-dotenv==1.0.1
loguru==0.7.2
alembic==1.13.2
Mako==1.3.5

# Data & Utils
pandas==2.2.2
numpy==1.26.4
requests==2.32.2
aiohttp==3.9.5
ta==0.11.0
pandas_ta==0.3.14b0
python-dateutil==2.9.0.post0

# Market Data APIs
alpha-vantage==3.0.0
yfinance==0.2.37
finnhub-python==2.4.20
twelvedata==1.2.17
fmpsdk==20250102.0

# Broker SDKs
ib_insync==0.9.81
tastytrade==10.3.0

# Discord / Notifications
discord-webhook==1.3.0
schedule==1.2.1

# Task queue / async
celery==5.3.6
redis==5.0.4

# Linters & Dev
black==24.4.2
ruff==0.5.5

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6

# Auth deps
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
